Revert "Merge pull request #392 from agentclash/worktree-lazy-growing-peach"
This reverts commit 409d2fe808e7d2172643e648826fda9fd26f62f9, reversing
changes made to b3c76932ab71b4cd69bcd6acb32810e879bc4405.
