+++ web/src/app/(workspace)/workspaces/[workspaceId]/runs/create-run-dialog.tsx
@@ -27,30 +27,18 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import { GlossaryTerm } from "@/components/onboarding/glossary-term";
 import { toast } from "sonner";
 import { Loader2, Plus } from "lucide-react";
 
 interface CreateRunDialogProps {
   workspaceId: string;
-  open?: boolean;
-  onOpenChange?: (open: boolean) => void;
 }
 
-export function CreateRunDialog({
-  workspaceId,
-  open: controlledOpen,
-  onOpenChange: controlledOnOpenChange,
-}: CreateRunDialogProps) {
+export function CreateRunDialog({ workspaceId }: CreateRunDialogProps) {
   const router = useRouter();
   const { getAccessToken } = useAccessToken();
 
-  const [internalOpen, setInternalOpen] = useState(false);
-  const open = controlledOpen ?? internalOpen;
-  const setOpen = (next: boolean) => {
-    if (controlledOpen === undefined) setInternalOpen(next);
-    controlledOnOpenChange?.(next);
-  };
+  const [open, setOpen] = useState(false);
   const [name, setName] = useState("");
   const [selectedPackId, setSelectedPackId] = useState("");
   const [selectedVersionId, setSelectedVersionId] = useState("");
@@ -390,7 +378,7 @@ export function CreateRunDialog({
           {/* Challenge Pack */}
           <div>
             <label className="mb-1.5 block text-sm font-medium">
-              <GlossaryTerm term="challenge-pack">Challenge Pack</GlossaryTerm>
+              Challenge Pack
             </label>
             <select
               aria-label="Challenge Pack"
@@ -441,7 +429,7 @@ export function CreateRunDialog({
           {/* Input Set */}
           <div>
             <label className="mb-1.5 block text-sm font-medium">
-              <GlossaryTerm term="input-set">Input Set</GlossaryTerm>
+              Input Set
             </label>
             {!selectedVersionId ? (
               <p className="text-sm text-muted-foreground">
@@ -493,11 +481,7 @@ export function CreateRunDialog({
 
           <div className="space-y-3 rounded-lg border border-border bg-muted/20 p-3">
             <div>
-              <p className="text-sm font-medium">
-                <GlossaryTerm term="regression-coverage">
-                  Regression Coverage
-                </GlossaryTerm>
-              </p>
+              <p className="text-sm font-medium">Regression Coverage</p>
               <p className="text-xs text-muted-foreground">
                 Optionally add regression suites or specific cases to this run.
               </p>
@@ -583,9 +567,7 @@ export function CreateRunDialog({
 
                 <div>
                   <label className="mb-1.5 block text-sm font-medium">
-                    <GlossaryTerm term="official-pack-mode">
-                      Official Pack Mode
-                    </GlossaryTerm>
+                    Official Pack Mode
                   </label>
                   <select
                     aria-label="Official Pack Mode"
@@ -611,7 +593,7 @@ export function CreateRunDialog({
           {/* Agent Deployments (multi-select) */}
           <div>
             <label className="mb-1.5 block text-sm font-medium">
-              <GlossaryTerm term="deployment">Agent Deployments</GlossaryTerm>
+              Agent Deployments
             </label>
             {loading ? (
               <p className="text-sm text-muted-foreground">Loading...</p>
