+++ web/src/components/app-shell/top-bar.tsx
@@ -95,7 +95,6 @@ export function TopBar({
           avatarUrl={avatarUrl}
           orgName={orgName}
           orgSlug={orgSlug}
-          workspaceId={workspaceId}
         />
       </div>
     </header>
